chore(service): refactor service package files (#141)
Move firewall, server, IP address and tag interfaces and implementations to their own files.
