+++ upcloud/service/service.go
@@ -3,63 +3,14 @@ package service
 import (
 	"encoding/json"
 	"fmt"
-	"time"
 
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/client"
-	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/request"
 )
 
 type requestable interface {
 	RequestURL() string
 }
-
-type Firewall interface {
-	GetFirewallRules(r *request.GetFirewallRulesRequest) (*upcloud.FirewallRules, error)
-	GetFirewallRuleDetails(r *request.GetFirewallRuleDetailsRequest) (*upcloud.FirewallRule, error)
-	CreateFirewallRule(r *request.CreateFirewallRuleRequest) (*upcloud.FirewallRule, error)
-	CreateFirewallRules(r *request.CreateFirewallRulesRequest) error
-	DeleteFirewallRule(r *request.DeleteFirewallRuleRequest) error
-}
-
-var _ Firewall = (*Service)(nil)
-
-type IpAddress interface {
-	GetIPAddresses() (*upcloud.IPAddresses, error)
-	GetIPAddressDetails(r *request.GetIPAddressDetailsRequest) (*upcloud.IPAddress, error)
-	AssignIPAddress(r *request.AssignIPAddressRequest) (*upcloud.IPAddress, error)
-	ModifyIPAddress(r *request.ModifyIPAddressRequest) (*upcloud.IPAddress, error)
-	ReleaseIPAddress(r *request.ReleaseIPAddressRequest) error
-}
-
-var _ IpAddress = (*Service)(nil)
-
-type Server interface {
-	GetServerConfigurations() (*upcloud.ServerConfigurations, error)
-	GetServers() (*upcloud.Servers, error)
-	GetServerDetails(r *request.GetServerDetailsRequest) (*upcloud.ServerDetails, error)
-	CreateServer(r *request.CreateServerRequest) (*upcloud.ServerDetails, error)
-	WaitForServerState(r *request.WaitForServerStateRequest) (*upcloud.ServerDetails, error)
-	StartServer(r *request.StartServerRequest) (*upcloud.ServerDetails, error)
-	StopServer(r *request.StopServerRequest) (*upcloud.ServerDetails, error)
-	RestartServer(r *request.RestartServerRequest) (*upcloud.ServerDetails, error)
-	ModifyServer(r *request.ModifyServerRequest) (*upcloud.ServerDetails, error)
-	DeleteServer(r *request.DeleteServerRequest) error
-	DeleteServerAndStorages(r *request.DeleteServerAndStoragesRequest) error
-}
-
-var _ Server = (*Service)(nil)
-
-type Tag interface {
-	GetTags() (*upcloud.Tags, error)
-	CreateTag(r *request.CreateTagRequest) (*upcloud.Tag, error)
-	ModifyTag(r *request.ModifyTagRequest) (*upcloud.Tag, error)
-	DeleteTag(r *request.DeleteTagRequest) error
-	TagServer(r *request.TagServerRequest) (*upcloud.ServerDetails, error)
-	UntagServer(r *request.UntagServerRequest) (*upcloud.ServerDetails, error)
-}
-
-var _ Tag = (*Service)(nil)
 
 type PriceZones interface {
 	GetPriceZones() (*upcloud.PriceZones, error)
@@ -158,487 +109,6 @@ func (s *Service) GetPlans() (*upcloud.Plans, error) {
 	}
 
 	return &plans, nil
-}
-
-// GetServerConfigurations returns the available pre-configured server configurations
-func (s *Service) GetServerConfigurations() (*upcloud.ServerConfigurations, error) {
-	serverConfigurations := upcloud.ServerConfigurations{}
-	response, err := s.basicGetRequest("/server_size")
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &serverConfigurations)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverConfigurations, nil
-}
-
-// GetServers returns the available servers
-func (s *Service) GetServers() (*upcloud.Servers, error) {
-	servers := upcloud.Servers{}
-	response, err := s.basicGetRequest("/server")
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &servers)
-	if err != nil {
-		return nil, fmt.Errorf("unable to unmarshal JSON: %s, %w", string(response), err)
-	}
-
-	return &servers, nil
-}
-
-// GetServerDetails returns extended details about the specified server
-func (s *Service) GetServerDetails(r *request.GetServerDetailsRequest) (*upcloud.ServerDetails, error) {
-	serverDetails := upcloud.ServerDetails{}
-	response, err := s.basicGetRequest(r.RequestURL())
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, fmt.Errorf("unable to unmarshal JSON: %s, %w", string(response), err)
-	}
-
-	return &serverDetails, nil
-}
-
-// CreateServer creates a server and returns the server details for the newly created server
-func (s *Service) CreateServer(r *request.CreateServerRequest) (*upcloud.ServerDetails, error) {
-	serverDetails := upcloud.ServerDetails{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// WaitForServerState blocks execution until the specified server has entered the specified state. If the state changes
-// favorably, the new server details are returned. The method will give up after the specified timeout
-func (s *Service) WaitForServerState(r *request.WaitForServerStateRequest) (*upcloud.ServerDetails, error) {
-	attempts := 0
-	sleepDuration := time.Second * 5
-
-	for {
-		// Always wait for one attempt period before querying the state the first time. Newly created servers
-		// may not immediately switch to "maintenance" upon creation, triggering a false positive from this
-		// method
-		attempts++
-		time.Sleep(sleepDuration)
-
-		serverDetails, err := s.GetServerDetails(&request.GetServerDetailsRequest{
-			UUID: r.UUID,
-		})
-
-		if err != nil {
-			return nil, err
-		}
-
-		// Either wait for the server to enter the desired state or wait for it to leave the undesired state
-		if r.DesiredState != "" && serverDetails.State == r.DesiredState {
-			return serverDetails, nil
-		} else if r.UndesiredState != "" && serverDetails.State != r.UndesiredState {
-			return serverDetails, nil
-		}
-
-		if time.Duration(attempts)*sleepDuration >= r.Timeout {
-			return nil, fmt.Errorf("timeout reached while waiting for server to enter state \"%s\"", r.DesiredState)
-		}
-	}
-}
-
-// StartServer starts the specified server
-func (s *Service) StartServer(r *request.StartServerRequest) (*upcloud.ServerDetails, error) {
-	// Save previous timeout
-	prevTimeout := s.client.GetTimeout()
-
-	// Increase the client timeout to match the request timeout
-	s.client.SetTimeout(r.Timeout)
-
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	// Restore previous timout
-	s.client.SetTimeout(prevTimeout)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	serverDetails := upcloud.ServerDetails{}
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// StopServer stops the specified server
-func (s *Service) StopServer(r *request.StopServerRequest) (*upcloud.ServerDetails, error) {
-	// Save previous timeout
-	prevTimeout := s.client.GetTimeout()
-
-	// Increase the client timeout to match the request timeout
-	// Allow ten seconds to give the API a chance to respond with an error
-	s.client.SetTimeout(r.Timeout + 10*time.Second)
-
-	serverDetails := upcloud.ServerDetails{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	// Restore previous timeout
-	s.client.SetTimeout(prevTimeout)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// RestartServer restarts the specified server
-func (s *Service) RestartServer(r *request.RestartServerRequest) (*upcloud.ServerDetails, error) {
-	// Save previous timeout
-	prevTimeout := s.client.GetTimeout()
-
-	// Increase the client timeout to match the request timeout
-	// Allow ten seconds to give the API a chance to respond with an error
-	s.client.SetTimeout(r.Timeout + 10*time.Second)
-
-	serverDetails := upcloud.ServerDetails{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	// Restore previous timeout
-	s.client.SetTimeout(prevTimeout)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// ModifyServer modifies the configuration of an existing server. Attaching and detaching storages as well as assigning
-// and releasing IP addresses have their own separate operations.
-func (s *Service) ModifyServer(r *request.ModifyServerRequest) (*upcloud.ServerDetails, error) {
-	serverDetails := upcloud.ServerDetails{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPutRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// DeleteServer deletes the specified server
-func (s *Service) DeleteServer(r *request.DeleteServerRequest) error {
-	err := s.client.PerformJSONDeleteRequest(s.client.CreateRequestURL(r.RequestURL()))
-
-	if err != nil {
-		return parseJSONServiceError(err)
-	}
-
-	return nil
-}
-
-// DeleteServerAndStorages deletes the specified server and all attached storages
-func (s *Service) DeleteServerAndStorages(r *request.DeleteServerAndStoragesRequest) error {
-	err := s.client.PerformJSONDeleteRequest(s.client.CreateRequestURL(r.RequestURL()))
-
-	if err != nil {
-		return parseJSONServiceError(err)
-	}
-
-	return nil
-}
-
-// TagServer tags a server with with one or more tags
-func (s *Service) TagServer(r *request.TagServerRequest) (*upcloud.ServerDetails, error) {
-	serverDetails := upcloud.ServerDetails{}
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), nil)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// UntagServer removes one or more tags from a server
-func (s *Service) UntagServer(r *request.UntagServerRequest) (*upcloud.ServerDetails, error) {
-	serverDetails := upcloud.ServerDetails{}
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), nil)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &serverDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &serverDetails, nil
-}
-
-// CreateTag creates a new tag, optionally assigning it to one or more servers at the same time
-func (s *Service) CreateTag(r *request.CreateTagRequest) (*upcloud.Tag, error) {
-	tagDetails := upcloud.Tag{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &tagDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &tagDetails, nil
-}
-
-// ModifyTag modifies a tag (e.g. renaming it)
-func (s *Service) ModifyTag(r *request.ModifyTagRequest) (*upcloud.Tag, error) {
-	tagDetails := upcloud.Tag{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPutRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &tagDetails)
-	if err != nil {
-		return nil, err
-	}
-
-	return &tagDetails, nil
-}
-
-// DeleteTag deletes the specified tag
-func (s *Service) DeleteTag(r *request.DeleteTagRequest) error {
-	err := s.client.PerformJSONDeleteRequest(s.client.CreateRequestURL(r.RequestURL()))
-
-	if err != nil {
-		return parseJSONServiceError(err)
-	}
-
-	return nil
-}
-
-// GetIPAddresses returns all IP addresses associated with the account
-func (s *Service) GetIPAddresses() (*upcloud.IPAddresses, error) {
-	ipAddresses := upcloud.IPAddresses{}
-	response, err := s.basicGetRequest("/ip_address")
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &ipAddresses)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ipAddresses, nil
-}
-
-// GetIPAddressDetails returns extended details about the specified IP address
-func (s *Service) GetIPAddressDetails(r *request.GetIPAddressDetailsRequest) (*upcloud.IPAddress, error) {
-	ipAddress := upcloud.IPAddress{}
-	response, err := s.basicGetRequest(r.RequestURL())
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &ipAddress)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ipAddress, nil
-}
-
-// AssignIPAddress assigns the specified IP address to the specified server
-func (s *Service) AssignIPAddress(r *request.AssignIPAddressRequest) (*upcloud.IPAddress, error) {
-	ipAddress := upcloud.IPAddress{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &ipAddress)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ipAddress, nil
-}
-
-// ModifyIPAddress modifies the specified IP address
-func (s *Service) ModifyIPAddress(r *request.ModifyIPAddressRequest) (*upcloud.IPAddress, error) {
-	ipAddress := upcloud.IPAddress{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPatchRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &ipAddress)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ipAddress, nil
-}
-
-// ReleaseIPAddress releases the specified IP address from the server it is attached to
-func (s *Service) ReleaseIPAddress(r *request.ReleaseIPAddressRequest) error {
-	err := s.client.PerformJSONDeleteRequest(s.client.CreateRequestURL(r.RequestURL()))
-
-	if err != nil {
-		return parseJSONServiceError(err)
-	}
-
-	return nil
-}
-
-// GetFirewallRules returns the firewall rules for the specified server
-func (s *Service) GetFirewallRules(r *request.GetFirewallRulesRequest) (*upcloud.FirewallRules, error) {
-	firewallRules := upcloud.FirewallRules{}
-	response, err := s.basicGetRequest(r.RequestURL())
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &firewallRules)
-	if err != nil {
-		return nil, err
-	}
-
-	return &firewallRules, nil
-}
-
-// GetFirewallRuleDetails returns extended details about the specified firewall rule
-func (s *Service) GetFirewallRuleDetails(r *request.GetFirewallRuleDetailsRequest) (*upcloud.FirewallRule, error) {
-	firewallRule := upcloud.FirewallRule{}
-	response, err := s.basicGetRequest(r.RequestURL())
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &firewallRule)
-	if err != nil {
-		return nil, err
-	}
-
-	return &firewallRule, nil
-}
-
-// CreateFirewallRule creates the firewall rule
-func (s *Service) CreateFirewallRule(r *request.CreateFirewallRuleRequest) (*upcloud.FirewallRule, error) {
-	firewallRule := upcloud.FirewallRule{}
-	requestBody, _ := json.Marshal(r)
-	response, err := s.client.PerformJSONPostRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return nil, parseJSONServiceError(err)
-	}
-
-	err = json.Unmarshal(response, &firewallRule)
-	if err != nil {
-		return nil, err
-	}
-
-	return &firewallRule, nil
-}
-
-// CreateFirewallRules creates multiple firewall rules
-func (s *Service) CreateFirewallRules(r *request.CreateFirewallRulesRequest) error {
-	requestBody, _ := json.Marshal(r)
-	_, err := s.client.PerformJSONPutRequest(s.client.CreateRequestURL(r.RequestURL()), requestBody)
-
-	if err != nil {
-		return parseJSONServiceError(err)
-	}
-
-	return nil
-}
-
-// DeleteFirewallRule deletes the specified firewall rule
-func (s *Service) DeleteFirewallRule(r *request.DeleteFirewallRuleRequest) error {
-	err := s.client.PerformJSONDeleteRequest(s.client.CreateRequestURL(r.RequestURL()))
-
-	if err != nil {
-		return parseJSONServiceError(err)
-	}
-
-	return nil
-}
-
-// GetTags returns all tags
-func (s *Service) GetTags() (*upcloud.Tags, error) {
-	tags := upcloud.Tags{}
-	response, err := s.basicGetRequest("/tag")
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = json.Unmarshal(response, &tags)
-	if err != nil {
-		return nil, err
-	}
-
-	return &tags, nil
 }
 
 // Wrapper that performs a GET request to the specified location and returns the response or a service error
